+++ apps/api/src/domains/apps/apps.module.ts
@@ -15,6 +15,7 @@ import { AppsController } from "./apps.controller"
 import { AppsService } from "./apps.service"
 import { AppsInstallController } from "./apps-install.controller"
 import { AppsV1Controller } from "./apps-v1.controller"
+import { AppsDocumentSourcesController } from "./document-sources/apps-document-sources.controller"
 
 @Module({
   imports: [
@@ -24,7 +25,12 @@ import { AppsV1Controller } from "./apps-v1.controller"
     RbacModule,
     DocumentsModule,
   ],
-  controllers: [AppsController, AppsInstallController, AppsV1Controller],
+  controllers: [
+    AppsController,
+    AppsInstallController,
+    AppsV1Controller,
+    AppsDocumentSourcesController,
+  ],
   providers: [
     AppsService,
     AppManifestRepository,